Move ClimateSettingSchedule object to its own file
